Create README.md
This is a project based on the front end as well as the backend. Here you can track the live data of coronavirus in India. I use HTML, CSS, and JAVA-SCRIPT as Frontend and PHP as backend development. and I also use My SQL database for storing the data of users and live API for the real coronavirus update in our country.